Remove AmortizedMemoryStore implementation
- Removed AmortizedMemoryStore due to HashMap iteration order concerns
- Moved ProbabilisticMemoryStore to its own module (probabilistic.rs)
- Updated all imports, tests, benchmarks, and examples
- Now have 4 high-quality store implementations:
  - AdaptiveMemoryStore (recommended default)
  - OptimizedMemoryStore (simple and reliable)
  - InternedMemoryStore (for repeated keys)
  - ProbabilisticMemoryStore (low overhead cleanup)
