+++ benches/cleanup_strategies.rs
@@ -2,8 +2,8 @@ use criterion::{Criterion, Throughput, criterion_group, criterion_main};
 use std::hint::black_box;
 use std::time::{Duration, SystemTime};
 use throttlecrab::core::store::adaptive_cleanup::AdaptiveMemoryStore;
-use throttlecrab::core::store::amortized::{AmortizedMemoryStore, ProbabilisticMemoryStore};
 use throttlecrab::core::store::optimized::OptimizedMemoryStore;
+use throttlecrab::core::store::probabilistic::ProbabilisticMemoryStore;
 use throttlecrab::{MemoryStore, RateLimiter};
 
 /// Benchmark cleanup strategies with expired entries
@@ -118,41 +118,6 @@ fn benchmark_with_expiration(c: &mut Criterion) {
         });
     });
 
-    group.bench_function("amortized_50pct_expired", |b| {
-        let mut limiter = RateLimiter::new(AmortizedMemoryStore::with_capacity(10_000));
-        let start_time = SystemTime::now();
-        let mut counter = 0u64;
-
-        // Pre-populate with mixed TTLs
-        for i in 0..10_000 {
-            let key = format!("key_{i}");
-            let ttl = if i % 2 == 0 { 1 } else { 3600 };
-            limiter
-                .rate_limit(&key, 100, 1000, ttl, 1, start_time)
-                .unwrap();
-        }
-
-        let bench_time = start_time + Duration::from_secs(2);
-
-        b.iter(|| {
-            let key = format!("bench_key_{}", counter % 1000);
-            counter += 1;
-
-            let (allowed, _result) = limiter
-                .rate_limit(
-                    black_box(&key),
-                    black_box(100),
-                    black_box(1000),
-                    black_box(60),
-                    black_box(1),
-                    black_box(bench_time),
-                )
-                .unwrap();
-
-            black_box(allowed)
-        });
-    });
-
     group.bench_function("probabilistic_50pct_expired", |b| {
         let mut limiter = RateLimiter::new(ProbabilisticMemoryStore::with_capacity(10_000));
         let start_time = SystemTime::now();
@@ -207,30 +172,6 @@ fn benchmark_latency_percentiles(c: &mut Criterion) {
             counter += 1;
 
             // Add with short TTL to force cleanups
-            let (allowed, _result) = limiter
-                .rate_limit(
-                    black_box(&key),
-                    black_box(10),
-                    black_box(100),
-                    black_box(1), // 1 second TTL
-                    black_box(1),
-                    black_box(SystemTime::now()),
-                )
-                .unwrap();
-
-            black_box(allowed)
-        });
-    });
-
-    // Amortized - should have consistent latency
-    group.bench_function("amortized_consistent", |b| {
-        let mut limiter = RateLimiter::new(AmortizedMemoryStore::new());
-        let mut counter = 0u64;
-
-        b.iter(|| {
-            let key = format!("key_{counter}");
-            counter += 1;
-
             let (allowed, _result) = limiter
                 .rate_limit(
                     black_box(&key),
