+++ examples/capacity_test.rs
@@ -1,6 +1,6 @@
 use std::time::SystemTime;
 use throttlecrab::RateLimiter;
-use throttlecrab::core::store::{amortized::AmortizedMemoryStore, optimized::OptimizedMemoryStore};
+use throttlecrab::core::store::optimized::OptimizedMemoryStore;
 
 fn test_store_capacity<S: throttlecrab::core::store::Store>(
     name: &str,
@@ -46,13 +46,6 @@ fn main() {
         test_store_capacity(
             "Optimized MemoryStore",
             RateLimiter::new(OptimizedMemoryStore::with_capacity(size / 2)), // Under-provision
-            size,
-        );
-
-        // Amortized store - should handle any size
-        test_store_capacity(
-            "Amortized MemoryStore",
-            RateLimiter::new(AmortizedMemoryStore::with_capacity(size / 2)), // Under-provision
             size,
         );
     }
